dependency updates, jdk13+ compatibility, clean-ups

+++ src/main/java/fjord/ast/Operator.java
@@ -2,14 +2,14 @@ package fjord.ast;
 
 public class Operator {
 
-  private final int op;
+    private final int op;
 
-  public Operator(char op) {
-    this.op = (int) op;
-  }
+    public Operator(char op) {
+        this.op = op;
+    }
 
-  public int op() {
-    return op;
-  }
+    public int op() {
+        return op;
+    }
 
 }
